app/requirements.txt
django
django-extensions
whitenoise

djangorestframework

elastic-apm
